+++ src/lexer/lexer.c
@@ -1,46 +1,5 @@
 #include "minishell.h"
 
-static bool	is_pipeline_token_valid(t_AST_type type[], t_dict *env)
-{
-	if (type[1] == PIPELINE && (type[0] != WORD || type[2] == PIPELINE))
-	{
-		error_with_exitcode(
-			(char *[]){BRED, MINISHELL, SYNTAX_ERROR, " `|'\n", END, NULL},
-			env, 2);
-		return (false);
-	}
-	if (type[1] == PIPELINE && type[2] == TYPE_END)
-	{
-		error_with_exitcode(
-			(char *[]){
-			BRED, MINISHELL, SYNTAX_ERROR, " `|'\n",
-			MULTILINE_ERROR, END, NULL}, env, 2);
-		return (false);
-	}
-	return (true);
-}
-
-static bool	is_redirect_token_valid(t_AST_type type[], char *str, t_dict *env)
-{
-	if (type[1] == REDIRECT && type[0] == REDIRECT)
-	{
-		error_with_exitcode(
-			(char *[]){
-			BRED, MINISHELL, SYNTAX_ERROR, " `", str, "'\n", END, NULL},
-			env, 2);
-		return (false);
-	}
-	if ((type[0] == REDIRECT && type[1] == TYPE_END)
-		|| (type[1] == REDIRECT && type[2] == TYPE_END))
-	{
-		error_with_exitcode(
-			(char *[]){
-			BRED, MINISHELL, SYNTAX_ERROR, " `newline'\n", END, NULL}, env, 2);
-		return (false);
-	}
-	return (true);
-}
-
 static void	type_arr_init(t_AST_type type[3], t_token tokens[], int i)
 {
 	if (!i)
@@ -70,7 +29,7 @@ static bool	is_tokens_valid(t_token tokens[], t_dict *env)
 	return (true);
 }
 
-t_token	*lexer(char *line, t_dict *env)
+t_token	*new_tokens_from_line(char *line, t_dict *env)
 {
 	t_list	*scan_list;
 	t_token	*tokens;
@@ -81,7 +40,7 @@ t_token	*lexer(char *line, t_dict *env)
 		del_list(&scan_list, del_scan_node);
 	else
 	{
-		tokens = tokenizer(scan_list);
+		tokens = new_tokens_from_list(scan_list);
 		if (!tokens)
 		{
 			env_set_exitcode(env, EXIT_FAILURE);
